Quiet some -Wconversion warnings

+++ plugins/sudoers/gram.c
@@ -2253,7 +2253,7 @@ yyreduce:
 				sudoerserror(N_("\"CHROOT\" is deprecated"));
 				YYERROR;
 			    } else if (parser_conf.verbose > 0) {
-				parser_warnx(parsed_policy.ctx, sudoers, this_lineno, sudolinebuf.toke_start + 1, false, false, N_("\"CHROOT\" is deprecated"));
+				parser_warnx(parsed_policy.ctx, sudoers, this_lineno, (int)(sudolinebuf.toke_start + 1), false, false, N_("\"CHROOT\" is deprecated"));
 			    }
 			    (yyval.string) = (yyvsp[0].string);
 			}
